+++ examples/counter2.rs
@@ -6,7 +6,7 @@ pub enum Message {
     Decrement,
 }
 
-fn main() {
+fn main() -> Result<(), Box<dyn std::error::Error>> {
     let app = app::App::default();
     let mut wind = Window::new(200, 200, 160, 200, "Counter");
     let mut frame = Frame::new(30, 80, 100, 40, "0");
@@ -20,8 +20,8 @@ fn main() {
     but_inc.emit(s, Message::Increment);
     but_dec.emit(s, Message::Decrement);
 
-    while app.wait().unwrap() {
-        let label: i32 = frame.label().parse().unwrap();
+    while app.wait()? {
+        let label: i32 = frame.label().parse()?;
         
         match r.recv() {
             Some(Message::Increment) => frame.set_label(&(label + 1).to_string()),
@@ -29,4 +29,5 @@ fn main() {
             None => (),
         }
     }
+    Ok(())
 }