+++ examples/image.rs
@@ -47,14 +47,15 @@
 // // This is an example showing using the inbuilt image functions
 
 use fltk::{app::*, frame::*, image::*, window::*};
+use std::error::Error;
 use std::path::PathBuf;
 
-fn main() {
+fn main() -> Result<(), Box<dyn Error>> {
     let app = App::default().set_scheme(AppScheme::Gleam);
     let mut wind = Window::new(100, 100, 400, 300, "Hello from rust");
     let mut frame = Frame::new(0, 0, 400, 300, "");
 
-    let mut image = SharedImage::load(&PathBuf::from("screenshots/calc.jpg")).unwrap();
+    let mut image = SharedImage::load(&PathBuf::from("screenshots/cailc.jpg"))?;
     image.scale(200, 200, true, true);
 
     frame.set_image(&image);
@@ -62,5 +63,6 @@ fn main() {
     wind.make_resizable(true);
     wind.show();
 
-    app.run().unwrap();
+    app.run()?;
+    Ok(())
 }