+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Systematically return Error when FLTK (C++) returns Error code.
 - Add emit method for widgets which take a app::Sender and a message to allow for message passing and shorted callback function. 
 - Added the rest of the horizontal valuator widgets.
+- Methods return raw pointers are marked as unsafe.
+- Impl Error for FltkError.
 
 ## [0.4.13] - 2020-05-25
 ### Changes